restructuring Schema / Decoder / TypeContext
