+++ Assets/Colyseus/Runtime/Colyseus/Serializer/NoneSerializer.cs
@@ -1,9 +1,12 @@
 namespace Colyseus
 {
+    // TODO: remove dummy state dependency from NoneSerializer.
+    public class NoState : Schema.Schema { }
+
     /// <summary>
     ///     An empty implementation of <see cref="IColyseusSerializer{T}" />
     /// </summary>
-    public class ColyseusNoneSerializer : IColyseusSerializer<object>
+    public class NoneSerializer<T> : IColyseusSerializer<object> where T: NoState
     {
         /// <inheritdoc />
         public void SetState(byte[] rawEncodedState, int offset)
