--- Assets/Colyseus/Runtime/Colyseus/Serializer/Schema/Context.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: be586e40b2425468c9eb42d356a9cbb6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
