--- Assets/Colyseus/Runtime/Colyseus/Serializer/Schema/Encoder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6e4b415e783324279b8f0ef6d5f47821
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
