+++ Assets/Colyseus/Runtime/Colyseus/Serializer/Schema/StateCallbacks.cs
@@ -1,0 +1,18 @@
+
+namespace Colyseus.Schema
+{
+	public class Callbacks<T> where T : Schema
+	{
+		T state;
+
+		public Callbacks(T state)
+		{
+			this.state = state;
+		}
+
+		public static Callbacks<T> Get(T state)
+		{
+			return new Callbacks<T>(state);
+		}
+	}
+}
