+++ Assets/Colyseus/Runtime/Colyseus/Serializer/Schema/StateCallbacks.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cdc421113be204ebf86bc7061decb2a5
