+++ Assets/Colyseus/Runtime/Colyseus/Serializer/Schema/TypeContext.cs
@@ -5,28 +5,14 @@ namespace Colyseus.Schema
     /// <summary>
     ///     Internal container class for tracking and maintaining the context of the current application
     /// </summary>
-    internal class ColyseusContext
+    public class TypeContext
     {
-        /// <summary>
-        ///     Singleton instance
-        /// </summary>
-        protected static ColyseusContext instance = new ColyseusContext();
-
         /// <summary>
         ///     Schema types and IDs we're aware of
         /// </summary>
         protected Dictionary<float, System.Type> typeIds = new Dictionary<float, System.Type>();
 
         protected List<System.Type> types = new List<System.Type>(); //TODO: This does not appear to be used ever
-
-        /// <summary>
-        ///     Getter function for the singleton <see cref="instance" />
-        /// </summary>
-        /// <returns>The singleton <see cref="instance" /></returns>
-        public static ColyseusContext GetInstance()
-        {
-            return instance;
-        }
 
         /// <summary>
         ///     Add a new Schema type by id
