+++ Assets/Colyseus/Runtime/Colyseus/Serializer/Schema/TypeContext.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a1b0877a82fcf4867af631924169de16
