+++ Assets/Colyseus/Runtime/Colyseus/Serializer/Schema/Utils.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bb2c980af108447b0ae4b3af69848342
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
