+++ Assets/Colyseus/Runtime/Colyseus/Serializer/Schema/Utils/Decode.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: dd65c771c1d0b44399010708ca01d75d
