+++ Assets/Colyseus/Runtime/Colyseus/Serializer/Schema/Utils/Encode.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8b9a0de6846c54eec8dcc45f10ed3b1c
