+++ Packages/manifest.json
@@ -1,13 +1,13 @@
 {
   "dependencies": {
-    "com.unity.collab-proxy": "2.0.3",
-    "com.unity.ide.rider": "3.0.20",
-    "com.unity.ide.visualstudio": "2.0.18",
-    "com.unity.ide.vscode": "1.2.5",
-    "com.unity.test-framework": "2.0.1-exp.2",
-    "com.unity.textmeshpro": "3.0.6",
-    "com.unity.timeline": "1.6.4",
-    "com.unity.ugui": "1.0.0",
+    "com.unity.ai.navigation": "2.0.5",
+    "com.unity.collab-proxy": "2.6.0",
+    "com.unity.ide.rider": "3.0.31",
+    "com.unity.ide.visualstudio": "2.0.22",
+    "com.unity.multiplayer.center": "1.0.0",
+    "com.unity.timeline": "1.8.7",
+    "com.unity.ugui": "2.0.0",
+    "com.unity.modules.accessibility": "1.0.0",
     "com.unity.modules.ai": "1.0.0",
     "com.unity.modules.androidjni": "1.0.0",
     "com.unity.modules.animation": "1.0.0",
