+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2021.3.24f1
-m_EditorVersionWithRevision: 2021.3.24f1 (cf10dcf7010d)
+m_EditorVersion: 6000.0.32f1
+m_EditorVersionWithRevision: 6000.0.32f1 (b2e806cf271c)
